+++ changelog.md
@@ -37,6 +37,7 @@
 
 # Unreleased
 
+- Add /zm inventories - Allows access to the inventories of the online inventory editor
 - Add PlayerInteractEvent in DupeListener
 - Fix classic meta with PAPI color