+++ changelog.md
@@ -64,9 +64,10 @@
 
 - Add BlockPlaceEvent for Anti Dupe
 - Add placeholder %zmenu_test% for testing
-- Fix sound action, pitch and volume was reversed
 - Add ``job`` requirement. Check if the player has a job with JobsReborn plugins
 - Added a cache on placeholders. By default, the cache is 500ms. That corresponds to opening a menu. So if you have several times the same placeholder used, it will only be called once.
+- Fix sound action, pitch and volume was reversed
+- Fix commands async
 
 # 1.0.2.5